+++ ErsatzTV.Application/Channels/Queries/GetAllChannelsForSort.cs
@@ -1,0 +1,3 @@
+﻿namespace ErsatzTV.Application.Channels;
+
+public record GetAllChannelsForSort : IRequest<List<ChannelSortViewModel>>;